requirements.txt
annotated-types==0.7.0
anyio==4.0.0
click==8.1.8
fastapi==0.115.11
h11==0.14.0
idna==3.10
mysql-connector-python==9.0.0
pydantic==2.10.6
pydantic_core==2.27.2
python-dotenv==1.0.1
pytz==2025.1
sniffio==1.3.1
starlette==0.44.0
typing_extensions==4.12.2
uvicorn==0.33.0
